backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
chromadb
sentence-transformers
scikit-learn
numpy
google-generativeai
wikipedia
groq
streamlit
